requirements.txt
Flask==3.0.3
firebase-admin==6.5.0
flask-cors==4.0.1
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
numpy==1.24.3
sentence-transformers==2.2.2
